sched: ems: pcf: Optimize 'performance' CPU selection
The Performance CPU Finder (PCF) is pretty bare-bones, much more bare-bones than EMS' 'prefer-idle' CPU selection. Tasks apparently do get placed using this logic, even if the 'service' CPU selection already handles 'prefer-perf'. Let's improve this, just like all other EMS' selection paths.

Consider the utilization of the task on a given CPU before testing for any other placement, so we skip CPUs that will be over-capacity if the task is placed there. When selecting for cstates on idle cores, prefer the one with the lowest utilization if they are the same.

For non-idle CPUs, improve the spare capacity calculation to closely match CFS' and ensure that the backup CPU has the lowest wake util in addition to the active utilization. Tasks that are 'prefer-perf' might not necessarily be latency-sensitive, but they are performance-sensitive so ensure that tasks here get placed on (very) performant cores.

Signed-off-by: John Vincent <git@tenseventyseven.cf>